OPTIONAL: move ADSB.im instructions to that site
Feel free to drop this one before merging. I think it makes sense to
consolidate the instructions to live in one spot, but if you'd like to keep
your own version,I'll be happy to edit it here instead. It just is easier to
only have to maintain it once.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>